refactor(canon): split package route cache tests
